+++ src/features/user/MFV2/ContributionsMap/MyForestV2.module.scss
@@ -1,0 +1,12 @@
+.mapCreditsContainer {
+  position: absolute;
+  right: 48px;
+  bottom: 2px;
+  z-index: 1;
+  cursor: pointer;
+}
+
+.mapCredit {
+  padding: 10px;
+  width: 330px;
+}
